Development

See merge request megane_solliard/vseat!13

+++ VSEat_Project/BLL/OrderManager.cs
@@ -5,12 +5,16 @@ using System.Collections.Generic;
 
 namespace BLL
 {
-    public class OrderManager
+    public class OrderManager : IOrderManager
     {
 
         private IOrderDB OrderDb { get; }
         private IUserDB UserDb { get; }
         private IDelivererDB DelivererDb { get; }
+        private IOrderDetailDB OrderDetailDb { get; }
+        private ILocationDB LocationDb { get; }
+        private IMenuDB MenuDb { get; }
+        private IRestaurantDB RestaurantDb { get; }
 
         //Constructor
         public OrderManager(IConfiguration conf)
@@ -36,7 +40,7 @@ namespace BLL
             return OrderDb.GetOrdersForUser(userId);
         }
 
-        //Method to get one specific order
+        //Method to get one specific order with his ID
         public Order GetOrderWithID(int OrderID)
         {
             return OrderDb.GetOrderWithID(OrderID);
@@ -55,11 +59,30 @@ namespace BLL
         }
 
         //Add a new order
-        public Order AddOrder(Order order)
+        public Order AddOrder(User user, DateTime requiredDate, string shipAddress, string locationName)
         {
-            //Validation si le deliverer est dans la meme ville que le restaurant qui conerce l'ordre
-            //Augmenter le number Orders assigned dans le deliverer dès qu'on ajoute un order ayant son ID
-            return OrderDb.AddOrder(order);
+            int locationID = LocationDb.GetLocationWithName(locationName);
+
+            //Validation if User is connected
+            string isConnected = "Connected";
+            if (user.StatusAccount.Equals(isConnected))
+            {
+                //Add order informations
+                Order order = new Order(requiredDate, shipAddress, locationID);
+                order.UserID = user.UserID;
+                order.Price = UpdateOrderPrice(order, user);
+                //Ajout du deliverer ID
+                    //Validation if deliverer est dans la meme ville que le restaurant qui concerne l'ordre
+                    //with method checkCity(delivererID, restaurantID) dans deliverer manager
+                    //DelivererManager delivmnger = new DelivererManager(conf);
+
+                //Augmenter le number Orders assigned dans le deliverer dès qu'on ajoute un order ayant son ID
+                    //deliverer = DelivererDb.UpdateDeliverer(deliverer);
+
+                return OrderDb.AddOrder(order);
+            }
+            else
+                throw new BusinessExceptions("The user should be connected to be able to make an order");
         }
 
         //Update data for one order
@@ -69,41 +92,100 @@ namespace BLL
         }
 
         //Delete an order
-        public void DeleteOrder(int orderId, int userId)
+        public void DeleteOrder(int orderId, string firstname, string lastname)
         {
-            //var user = UserDb.GetUser(userId);
-            //var order = OrderDb.GetOrdersForUser(orderId,userId);
+            //Check that the firstname and the lastname correspond to the user who made the order
+            var userId = UserDb.GetUserIDWithName(firstname, lastname);
+            var user = UserDb.GetUserWithID(userId);
+            var order = OrderDb.GetOrderForUser(orderId, userId);
 
-            //if()
-            OrderDb.DeleteOrder(orderId, userId);
+            //current date must be at least 3 hours before shipped Date
+            var diffOfDates = order.RequiredDate - DateTime.Now; //get difference of two dates
+            Console.WriteLine("Difference in Hours: {0}", diffOfDates.Hours);
+
+            if (diffOfDates.Hours >= 3)
+            {
+                OrderDb.DeleteOrder(orderId, userId);
+            }
+            else
+            {
+                throw new BusinessExceptions("You cannot delete an order less than 3 hours before its delivery");
+            }
         }
+
+
         //Update RequiredDate for one order
         public String UpdateOrderRequiredDate(Order order,User user, DateTime newRequiredDate)
         {
-            string DateIsChanged = null;
 
-            Order orderChanged = null;
-            orderChanged = OrderDb.GetOrderForUser(order.OrderID, user.UserID);
+            //GESTION DES 15 MINUTES EN + !!!!
+            string DateIsChanged;
+
+            Order orderChanged = OrderDb.GetOrderForUser(order.OrderID, user.UserID);
+
             orderChanged.RequiredDate = newRequiredDate;
+
             orderChanged = OrderDb.UpdateOrder(orderChanged);
+
             DateIsChanged = "The Required Date has been changed !";
 
             return DateIsChanged;
         }
 
+        //Update RequiredDate for one order
+        public int UpdateOrderPrice(Order order, User user)
+        {
+            int price = 0;
+            //Reprendre tout les orders details ayant le meme orderID
+            List<OrderDetail> ordersDetails = OrderDetailDb.GetOrdersDetailsByOrder(order.OrderID);
+
+            //Adition of all total amount in order Details to form the price of the order
+            foreach (OrderDetail orderDet in ordersDetails)
+            {
+                price += orderDet.TotalAmount;
+            }
+
+            //Update in the DB
+            Order orderChanged = OrderDb.GetOrderForUser(order.OrderID, user.UserID);
+            orderChanged.Price = price;
+            orderChanged = OrderDb.UpdateOrder(orderChanged);
+
+            Console.WriteLine("The Price of the order is  " + orderChanged.Price + ".-");
+
+            return orderChanged.Price;
+        }
+
+
+        //Update Deliverer Id
+        public int UpdateDelivererID(Order order, Deliverer deliverer)
+        {
+            int delivererID = 0;
+            
+            //CheckCity
+            if(available = "yes")
+            {
+                order.DelivererID = deliverer.DelivererID;
+            }
+
+
+            return delivererID;
+        }
+
         public string OrderIsDelivered(Order order, Deliverer deliverer)
         {
-            string isDelivered = null;
+            string isDelivered;
 
-            Order orderChanged = null;
-            orderChanged = OrderDb.GetOrderForUser(order.OrderID, deliverer.DelivererID);
+            Order orderChanged = OrderDb.GetOrderForDeliverer(order.OrderID, deliverer.DelivererID);
+            Deliverer delivererChanged = DelivererDb.GetDeliverer(deliverer.DelivererID);
 
             //Décrémenter le number orders assigned du Deliverer
-            //Modifier la shipdate en mettant la date courrant
+            delivererChanged.NumberOrdersAssigned--;
+
+            //Modifier la shipdate en mettant la date courrante
+            orderChanged.ShippedDate = DateTime.Now;
 
             //Modifier le status de l'ordre
             orderChanged.StatusOrder = "Shipped";
-
 
             orderChanged = OrderDb.UpdateOrder(orderChanged);
             isDelivered = "The Order has been shipped !";
@@ -111,9 +193,15 @@ namespace BLL
             return isDelivered;
         }
 
-        public void GetOrderLocation()
+        public Location GetLocationRestaurant(Order order)
         {
+            OrderDetail orderDetail = OrderDetailDb.GetOrderDetailWithOrderID(order.OrderID);
+            Menu menu = MenuDb.GetMenuWithID(orderDetail.MenuID);
 
+            Restaurant restaurant = RestaurantDb.GetRestaurantWithID(menu.RestaurantID);
+            Location location = LocationDb.GetLocationWithID(restaurant.LocationID);
+
+            return location;
         }
 
     }
